backend/requirements.txt
jira
langchain-community
langchain-openai
langgraph
lxml
flask
flask-cors
node
npm
